🚚 Rename: 리포트 취향 카드 폴더 이동
