+++ apps/web/src/features/report/index.ts
@@ -1,4 +1,4 @@
-export { default as ReportPreferenceCard } from './components/ReportPreferenceCard';
-export { default as ReportPreferenceShareCard } from './components/ReportPreferenceShareCard';
+export { default as ReportPreferenceCard } from './components/report-preference-card/ReportPreferenceCard';
+export { default as ReportPreferenceShareCard } from './components/report-preference-card/ReportPreferenceShareCard';
 export { useReportImageDownload } from './hooks/useReportImageDownload';
 export { MOCK_REPORT_PREFERENCE } from './mockData';
